+++ R/detect.R
@@ -1,43 +1,41 @@
-missInject <- function(TICdat,idx){
-    thresh <- quantile(TICdat$value)[2] - IQR(TICdat$value) * 1.5
-    
-    missinjections <- TICdat %>%
-        filter(value < thresh) %>%
-        select(idx) %>%
-        unlist() %>%
-        unname() %>%
-        list(idx = idx,missInjections = .)
-    return(missinjections)
-}
-
-#' detectMissInjections
+#' Detect miss injections
 #' @rdname detectMissInjections
-#' @description detect miss injected samples
-#' @param x object of class Binalysis or MetaboProfile
-#' @param idx info column to use for sample indexes
-#' @importFrom stats IQR quantile
+#' @description Detect miss injected samples.
+#' @param x object of class `Binalysis` or `MetaboProfile`
+#' @param idx sample information column to use for sample indexes
+#' @return A list containing the name of sample information column used to index the miss injections and a vector of miss injection indexes.
+#' @details 
+#' Samples with a total ion count (TIC) below 1.5 times the inter-quartile range are detected as miss injections.
 #' @examples 
-#' \dontrun{
-#' files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
+#' ## Retrieve file paths and sample information for example data
+#' files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]
 #' 
-#' info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')
+#' info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]
 #' 
+#' ## Perform spectral binning
 #' analysis <- binneR::binneRlyse(files, 
 #'                                info, 
 #'                                parameters = binneR::detectParameters(files))
 #' 
+#' ## Detect miss injections
 #' miss_injections <- detectMissInjections(analysis)
 #' 
+#' ## Display detected miss injections
 #' miss_injections$missInjections
-#' }
 #' @export
+
+setGeneric('detectMissInjections',function(x,idx = 'fileOrder')
+    standardGeneric('detectMissInjections'))
+
+#' @rdname detectMissInjections
+#' @importFrom tibble as_tibble
 
 setMethod('detectMissInjections',signature = 'Binalysis',
           function(x,idx = 'fileOrder'){
               
               i <- x %>%
                   binneR::sampleInfo() %>%
-                  select(idx)
+                  select(all_of(idx))
               
               x %>%
                   binnedData %>%
@@ -50,25 +48,26 @@ setMethod('detectMissInjections',signature = 'Binalysis',
           })
 
 #' @rdname detectMissInjections
-#' @export
+#' @importFrom profilePro technique
 
 setMethod('detectMissInjections',signature = 'MetaboProfile',
           function(x,idx = 'fileOrder'){
               
               i <- x %>%
-                  .@Info %>%
-                  select(idx)
+                  profilePro::sampleInfo() %>% 
+                  select(all_of(idx))
               
-              if (str_detect(x@processingParameters@technique,'GCMS')) {
-                  mi <- x %>%
-                      .@Data %>%
+              mi <- x %>% 
+                  processedData()
+              
+              if (str_detect(technique(x),'GCMS')) {
+                  mi <- mi %>% 
                       rowSums() %>%
                       tibble(value = .) %>%
                       bind_cols(i)
               } else {
-                  mi <- x %>%
-                      .@Data %>%
-                      map(rowSums) %>%
+                  mi <- mi %>% 
+                  map(rowSums) %>%
                       bind_cols() %>%
                       rowSums() %>%
                       as_tibble() %>%
@@ -78,10 +77,142 @@ setMethod('detectMissInjections',signature = 'MetaboProfile',
                   missInject(idx = idx) 
           })
 
+#' @importFrom stats IQR quantile
+
+missInject <- function(TICdat,idx){
+    thresh <- quantile(TICdat$value)[2] - IQR(TICdat$value) * 1.5
+    
+    missinjections <- TICdat %>%
+        filter(value < thresh) %>%
+        select(all_of(idx)) %>%
+        unlist() %>%
+        unname() %>%
+        list(idx = idx,missInjections = .)
+    return(missinjections)
+}
+
+
+#' Detect batch/block differences
+#' @rdname detectBatchDiff
+#' @description Detect batch/block differences within analytical runs for each ionisation mode.
+#' @param x object of class `Binalysis` or `MetaboProfile`
+#' @param by info class column to use for batch information
+#' @param pthresh p-value threshold for significance
+#' @return If no differences between batches are found then `NULL` is returned. If significant differences are found then a tibble is returned containing the ANOVA results for each ionisation mode and showing whether batch correction is needed. 
+#' @details Analysis of Variance (ANOVA) is used to detect differences in total ion count (TIC) averages between batches/blocks. 
+#' @examples 
+#' ## Retrieve file paths and sample information for example data
+#' files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]
+#' 
+#' info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]
+#' 
+#' ## Perform spectral binning
+#' analysis <- binneR::binneRlyse(files, 
+#'                                info, 
+#'                                parameters = binneR::detectParameters(files))
+#' 
+#' ## Detect batch differences
+#' batch_diff <- detectBatchDiff(analysis)
+#' 
+#' ## Display batch diffferences
+#' batch_diff
+#' @export
+
+setGeneric('detectBatchDiff',function(x, by = 'block', pthresh = 0.05)
+    standardGeneric('detectBatchDiff')
+)
+
+#' @rdname detectBatchDiff
+#' @importFrom binneR binnedData
+#' @importFrom tidyr gather
+#' @importFrom tibble rowid_to_column
+#' @importFrom dplyr group_by_all
+#' @importFrom tidyselect all_of
+
+setMethod('detectBatchDiff',signature = 'Binalysis',
+          function(x, by = 'block', pthresh = 0.05){
+              rawInfo <- x %>%
+                  binneR::sampleInfo()
+              
+              TICdat <- x %>%
+                  binnedData %>%
+                  map(rowSums) %>%
+                  bind_cols() %>%
+                  rowid_to_column(var = 'Sample') %>%
+                  mutate(batch = rawInfo[,by] %>% unlist() %>% factor()) %>%
+                  gather('Mode','TIC',-batch,-Sample)
+              
+              diff <- batchDiff(TICdat,pthresh)
+              return(diff)
+          }
+)
+
+#' @rdname detectBatchDiff
+#' @importFrom stringr str_detect
+#' @importFrom tibble deframe
+
+setMethod('detectBatchDiff',signature =  "MetaboProfile",
+          function(x, by = 'block', pthresh = 0.05){
+              ri <- x %>%
+                  profilePro::sampleInfo()
+              
+              TICdat <- x %>%
+                  processedData()
+              
+              if (str_detect(technique(x),'GCMS')) {
+                  TICdat <- TICdat %>%
+                      rowSums() %>%
+                      {tibble(Sample = 1:length(.),
+                              TIC = .,
+                              batch = ri[,by] %>% deframe() %>% factor())}
+              } else {
+                  TICdat <- TICdat %>% 
+                      map(rowSums) %>%
+                      bind_cols() %>%
+                      rowid_to_column(var = 'Sample') %>%
+                      mutate(batch = ri[,by] %>% unlist() %>% factor()) %>%
+                      gather('Mode','TIC',-batch,-Sample)   
+              }
+              
+              diff <- batchDiff(TICdat,pthresh)
+              return(diff) 
+          })
+
 #' @importFrom dplyr bind_rows mutate n
 #' @importFrom tibble tibble
 
 batchDiff <- function(TICdat,pthresh = 0.05){
+    
+    block_frequencies <- TICdat %>%
+        group_by(batch) %>%
+        summarise(n = n())
+    
+    if (nrow(block_frequencies) < 2) {
+        message('Only 1 batch detected, skipping detection')
+        return()
+    }
+    
+    if (TRUE %in% (block_frequencies$n < 3)) {
+        removal_blocks <- block_frequencies %>%
+            filter(n < 3)
+        
+        TICdat <- TICdat %>%
+            filter(!(batch %in% removal_blocks$batch))
+        
+        message(str_c('Batches with < 3 replicates removed: ',
+                      str_c(str_c('"',removal_blocks$batch,'"'),
+                            collapse = ', ')))
+        
+    }
+    
+    new_block_frequencies <- TICdat %>%
+        group_by(batch) %>%
+        summarise(n = n())
+    
+    if (nrow(new_block_frequencies) < 2) {
+        message('Only 1 batch detected, skipping detection')
+        return()
+    }
     
     if ('Mode' %in% colnames(TICdat)) {
         ANOVAres <- TICdat %>%
@@ -107,127 +238,206 @@ batchDiff <- function(TICdat,pthresh = 0.05){
     return(ANOVAres)
 }
 
-#' detectBatchDiff
-#' @description Detect batch differences
-#' @rdname detectBatchDiff
-#' @param x object of class Binalysis or MetaboProfile
-#' @param by info class column to use for batch information
-#' @param pthresh p-value threshold for significance
-#' @importClassesFrom binneR Binalysis
-#' @importFrom binneR binnedData
-#' @importFrom tidyr gather
-#' @importFrom tibble rowid_to_column
-#' @importFrom dplyr group_by_all
-#' @importFrom tidyselect all_of
+#' Detect pre-treatment parameters
+#' @rdname detectPretreatmentParameters
+#' @description Detect pre-treatment parameters for `Binalysis` or `MetaboProfile` class objects. 
+#' @param x S4 object of class `Binalysis`, `MetaboProfile` or `AnalysisData`
+#' @return S4 object of class `AnalysisParameters`
+#' @examples
+#' ## Retreive example file paths and sample information 
+#' file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes') %>% 
+#'    .[61:63]
+#' 
+#' sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes') %>% 
+#'     dplyr::filter(name == 'QC01' | name == 'QC02' | name == 'QC03')
+#' 
+#' ## Detect spectral binning parameters
+#' bp <- binneR::detectParameters(file_paths)
+#' 
+#' ## Perform spectral binning
+#' bd <- binneR::binneRlyse(file_paths,sample_information,bp)
+#' 
+#' ## Detect pre-treatment parameters
+#' pp <- detectPretreatmentParameters(bd) 
+#' 
+#' pp
 #' @export
 
-setMethod('detectBatchDiff',signature = 'Binalysis',
-          function(x, by = 'block', pthresh = 0.05){
-              rawInfo <- x %>%
-                  binneR::sampleInfo()
-              
-              TICdat <- x %>%
-                  binnedData %>%
-                  map(rowSums) %>%
-                  bind_cols() %>%
-                  rowid_to_column(var = 'Sample') %>%
-                  mutate(batch = rawInfo[,by] %>% unlist() %>% factor()) %>%
-                  gather('Mode','TIC',-batch,-Sample)
-              
-              i <- rawInfo %>%
-                  select(batch = all_of(by))
-              
-              clsFreq <- i %>%
-                  group_by_all() %>%
-                  summarise(n = n())
-              
-              if (T %in% (clsFreq$n < 3)) {
-                  clsRem <- clsFreq %>%
-                      filter(n < 3)
-                  
-                  TICdat <- TICdat %>%
-                      filter(!(batch %in% clsRem$batch))
-                  
-                  warning(str_c('Batches with < 3 replicates removed: ',str_c(str_c('"',clsRem$batch,'"'),collapse = ', ')),call. = F)
-                  
-              }
-              
-              diff <- batchDiff(TICdat,pthresh)
-              return(diff)
-          }
-)
+setGeneric('detectPretreatmentParameters',function(x){
+    standardGeneric('detectPretreatmentParameters')
+})
 
-#' @rdname detectBatchDiff
-#' @importFrom stringr str_detect
-#' @importFrom tibble deframe
-#' @export
+#' @rdname detectPretreatmentParameters
 
-setMethod('detectBatchDiff',signature =  "MetaboProfile",
-          function(x, by = 'block', pthresh = 0.05){
-              ri <- x %>%
-                  .@Info
-              
-              if (str_detect(x@processingParameters@technique,'GCMS')) {
-                  TICdat <- x %>%
-                      .@Data %>%
-                      rowSums() %>%
-                      {tibble(Sample = 1:length(.),
-                              TIC = .,
-                              batch = ri[,by] %>% deframe() %>% factor())}
-              } else {
-                  TICdat <- x %>%
-                      .@Data %>%
-                      map(rowSums) %>%
-                      bind_cols() %>%
-                      rowid_to_column(var = 'Sample') %>%
-                      mutate(batch = ri[,by] %>% unlist() %>% factor()) %>%
-                      gather('Mode','TIC',-batch,-Sample)   
-              }
-              
-              diff <- batchDiff(TICdat,pthresh)
-              return(diff) 
+setMethod('detectPretreatmentParameters',signature = 'Binalysis',
+          function(x){
+              detectPretreatment(x)
           })
 
-#' detectPairwises
-#' @description Detect availble pairwise comparisons
-#' @param x object of class Analysis
-#' @param cls info column to use for class information
-#' @param type type of analysis (classification or featureSelection)
-#' @importClassesFrom metabolyseR Analysis
-#' @importFrom utils combn
-#' @importFrom dplyr filter select
-#' @importFrom tibble as_tibble
-#' @importFrom purrr map_chr
-#' @importFrom stringr str_c
-#' @rdname detectPairwises
+#' @rdname detectPretreatmentParameters
+
+setMethod('detectPretreatmentParameters',signature = 'MetaboProfile',
+          function(x){
+              detectPretreatment(x)
+          })
+
+#' @importFrom metabolyseR parameters<- parameters
+
+detectPretreatment <- function(x){
+    miss_injections <- detectMissInjections(x)
+    batch_correction <- detectBatchDiff(x)
+    
+    
+    pre_treat_params <- analysisParameters('pre-treatment')
+    
+    if (TRUE %in% batch_correction$`Correction needed`) {
+        parameters(pre_treat_params,'pre-treatment') <- c(
+            list(
+                correction = list(
+                    center = list(block = 'block',
+                                  type = 'median'
+                    )),
+                parameters(pre_treat_params,'pre-treatment')
+            ))
+    }
+    
+    if (length(miss_injections$missInjections) > 0){
+        parameters(pre_treat_params,'pre-treatment') <- c(
+            list(
+                remove = list(
+                    samples = list(idx = miss_injections$idx,samples = miss_injections$missInjections)
+                )),
+            parameters(pre_treat_params)
+        )
+    }
+    
+    return(pre_treat_params)
+} 
+
+#' Detect modelling parameters
+#' @rdname detectModellingParameters
+#' @description Detect modelling parameters for `Binalysis`, `MetaboProfile` or `Analysis` S4 classes. 
+#' @param x S4 object of class `Binalysis`,`MetaboProfile` or `Analysis`
+#' @param type detect parameters for `raw` or `pre-treated` data for `Analysis` class
+#' @param cls sample information column to use for modelling
+#' @param ... arguments to pass to the appropriate method
+#' @return S4 object of class `AnalysisParameters`
+#' @examples 
+#' ## Retrieve file paths and sample information for example data
+#' files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]
+#' 
+#' info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]
+#' 
+#' ## Perform spectral binning
+#' analysis <- binneR::binneRlyse(files, 
+#'                                info, 
+#'                                parameters = binneR::detectParameters(files))
+#' 
+#' ## Detect modelling parameters
+#' modelling_parameters <- detectModellingParameters(analysis)
+#' 
+#' modelling_parameters
 #' @export
 
-setMethod('detectPairwises',signature = 'Analysis',
-          function(x,cls,type){
-              if (type == 'classification') {
-                  minimum <- 6
+setGeneric('detectModellingParameters',function(x,...){
+    standardGeneric('detectModellingParameters')
+})
+
+#' @rdname detectModellingParameters
+#' @importFrom binneR cls 
+
+setMethod('detectModellingParameters',signature = 'Binalysis',
+          function(x){
+              idx <- cls(x)
+              
+              if (length(idx) == 0){
+                  idx <- 'class'
               }
-              if (type == 'featureSelection') {
-                  minimum <- 3
-              }
-              info <- x %>%
-                  sinfo(type = 'pre-treated') %>% 
-                  select(!!cls) %>%
-                  unlist() %>%
-                  table() %>%
-                  as_tibble() %>%
-                  filter(n >= minimum) %>%
-                  select(1) %>%
-                  unlist() %>%
-                  sort()
-              if (length(info) > 0) {
-                  com <- combn(info,2) %>%
-                      t() %>%
-                      split(1:nrow(.)) %>%
-                      map_chr(str_c,collapse = '~')
-              } else {
-                  com <- character()
-              }
-              return(com)
-          }
-)
+              
+              sample_information <- binneR::sampleInfo(x) %>% 
+                  select(all_of(idx)) %>% 
+                  deframe()
+              
+              detectModelling(sample_information,idx)
+          })
+
+#' @rdname detectModellingParameters
+#' @importFrom profilePro processingParameters
+
+setMethod('detectModellingParameters',signature = 'MetaboProfile',
+          function(x){
+             idx <- x %>% 
+                 processingParameters() %>% 
+                 .$info %>% 
+                 .$cls
+             
+             sample_information <- x %>% 
+                 profilePro::sampleInfo() %>% 
+                 select(all_of(idx)) %>% 
+                 deframe()
+             
+             detectModelling(sample_information,idx)
+          })
+
+#' @rdname detectModellingParameters
+
+setMethod('detectModellingParameters',signature = 'Analysis',
+          function(x,type = 'pre-treated',cls = 'class'){
+              
+              sample_information <- x %>% 
+                  sinfo(type = type) %>% 
+                  select(all_of(cls)) %>% 
+                  deframe()
+              
+              detectModelling(sample_information,cls)
+          })
+
+#' @importFrom metabolyseR modellingParameters
+#' @importFrom dplyr filter
+
+detectModelling <- function(sample_info,idx){
+    
+    ## Detect if regression is needed
+    if (is.numeric(sample_info)){
+        message('Numeric class column detected, using regression')
+        if (length(sample_info) > 6){
+            detected_parameters <- modellingParameters('randomForest')
+        } else {
+            detected_parameters <- modellingParameters('linearRegression')
+        }
+        
+    } else {
+        ## Detect classification parameters
+        class_frequencies <- sample_info %>%
+            tibble(class = .) %>% 
+            group_by(class) %>% 
+            summarise(freq = n())
+        
+        if (nrow(class_frequencies) < 2){
+            detected_parameters <- list()
+        } else {
+            if (nrow(class_frequencies %>%
+                     filter(freq > 5)) < (floor(length(unique(sample_info)) / 2))) {
+                message('Less than 50% of classes have > 5 replicates. Using ANOVA.')
+                
+                detected_parameters <- modellingParameters('anova')
+            } else {
+                detected_parameters <- modellingParameters('randomForest')
+            }   
+        }
+        
+    }
+    
+    if (length(detected_parameters) > 0){
+        detected_parameters[[1]]$cls <- idx 
+        
+        if (names(detected_parameters)[1] == 'randomForest'){
+            detected_parameters[[1]]$reps <- 10
+        }
+    }
+    
+    modelling_parameters <- analysisParameters('modelling')
+    parameters(modelling_parameters,'modelling') <- detected_parameters
+    
+    return(modelling_parameters)
+}