+++ R/export.R
@@ -1,15 +1,22 @@
 #' Export results
 #' @rdname export
-#' @description Export data tables from analysis results.
-#' @param analysis S4 object of class Binalysis, Analysis, Assignment,
+#' @description Export data tables from `Binalysis`,`MetaboProfile`, `Analysis` and `Assignment` classes.
+#' @param analysis S4 object of class `Binalysis`, `MetaboProfile`, `Analysis` or `Assignment`,
 #' @param outPath directory path to export to.
+#' @export
+
+setGeneric('export',function(analysis,outPath = '.')
+    standardGeneric('export')
+)
+
+#' @rdname export
 #' @importFrom binneR binnedData
 #' @importFrom readr write_csv
 #' @importFrom purrr walk
 #' @importFrom stringr str_remove_all
-#' @export
 
-setMethod('export',signature = 'Binalysis',function(analysis,outPath = '.'){
+setMethod('export',signature = 'Binalysis',
+          function(analysis,outPath = '.'){
     exportPath <- str_c(outPath,'/exports')
     
     message('Exporting binned data')
@@ -21,7 +28,7 @@ setMethod('export',signature = 'Binalysis',function(analysis,outPath = '.'){
     i <- analysis %>%
         binneR::sampleInfo()
     
-    if (T %in% duplicated(i$name)) {
+    if (TRUE %in% duplicated(i$name)) {
         i <- fixNames(i)
     }
     
@@ -46,12 +53,52 @@ setMethod('export',signature = 'Binalysis',function(analysis,outPath = '.'){
 })
 
 #' @rdname export
+#' @importFrom profilePro peakInfo
+
+setMethod('export',signature = 'MetaboProfile',
+          function(analysis,outPath = '.'){
+              exportPath <- str_c(outPath,'/exports')
+              
+              message('Exporting spectrally processed data')
+              
+              if (!dir.exists(exportPath)) {
+                  dir.create(exportPath)
+              }
+              
+              i <- analysis %>%
+                  profilePro::sampleInfo()
+              
+              if (TRUE %in% duplicated(i$name)) {
+                  i <- fixNames(i)
+              }
+              
+              write_csv(i,str_c(exportPath,'/','raw_sample_info.csv'))
+              
+              write_csv(peakInfo(analysis) %>% 
+                            select(-peakidx),str_c(exportPath,'/peak_info.csv'))
+              
+              pd <- analysis %>%
+                  processedData() 
+              
+              names(pd)[is.na(names(pd))] <- '1'
+              
+              pd %>%
+                  names() %>%
+                  walk(~{
+                      bind_cols(pd[[.x]],i %>% select(name)) %>%
+                          gather('Feature','Intensity',-name) %>%
+                          spread(name,Intensity) %>%
+                          write_csv(str_c(exportPath,'/',.x,'_mode_raw_data.csv'))
+                  })
+          })
+
+#' @rdname export
 #' @importFrom dplyr bind_cols everything mutate select
 #' @importFrom tidyr gather spread
 #' @importFrom stringr str_sub str_split_fixed
-#' @export
 
-setMethod('export',signature = 'Analysis',function(analysis,outPath = '.'){
+setMethod('export',signature = 'Analysis',
+          function(analysis,outPath = '.'){
     exportPath <- str_c(outPath,'/exports')
     
     message('Exporting pre-treated data')
@@ -63,7 +110,7 @@ setMethod('export',signature = 'Analysis',function(analysis,outPath = '.'){
     i <- analysis %>%
         sinfo(type = 'pre-treated')
     
-    if (T %in% duplicated(i$name)) {
+    if (TRUE %in% duplicated(i$name)) {
         i <- fixNames(i)
     }
     
@@ -85,7 +132,6 @@ setMethod('export',signature = 'Analysis',function(analysis,outPath = '.'){
 
 #' @rdname export
 #' @importFrom MFassign assignments summariseAssignment
-#' @export
 
 setMethod('export',signature = 'Assignment',function(analysis,outPath = '.'){