+++ R/parallel.R
@@ -1,5 +1,5 @@
 #' Generate a suitable parallel processing plan
-#' @description Generate a suitable parallel processing future plan for your session. See ?future::plan for more details on future plans?
+#' @description Generate a suitable parallel processing future plan for your session. See `?future::plan` for more details on future plans.
 #' @examples 
 #' \dontrun{
 #' suitableParallelPlan()