+++ tests/testthat/test-parallel.R
@@ -1,0 +1,3 @@
+test_that("suitable parallel plan generated", {
+  expect_message(suitableParallelPlan())
+})