+++ tests/testthat/test-reduce.R
@@ -1,0 +1,9 @@
+test_that("reduce works", {
+    assigned_data <- metabolyseR::analysisData(
+        assignedData(assignment),
+        tibble::tibble(sample = seq_len(nrow(peakData)))
+    ) 
+    reduced_data <- reduce(assigned_data)
+    
+    expect_s4_class(reduced_data,'AnalysisData')
+})